+++ apps/trackers/jira/query.py
@@ -151,6 +151,12 @@ class TrackerJiraQueryBuilder(TrackerQueryBuilder):
         }:
             self._query["fields"]["labels"].append("validation-requested")
 
+        # If at least one affect has is_contract_priority, add label contract-priority
+        for affect in self.tracker.affects.all():
+            if affect.is_contract_priority:
+                self._query["fields"]["labels"].append("contract-priority")
+                break
+
     def generate_sla(self):
         """
         generate query for Jira SLA timestamps