+++ collectors/ps_constants/tasks.py
@@ -6,12 +6,18 @@ from celery.utils.log import get_task_logger
 from django.utils import timezone
 
 from collectors.framework.models import collector
-from osidb.models import CompliancePriority, SpecialConsiderationPackage, UbiPackage
+from osidb.models import (
+    CompliancePriority,
+    ContractPriority,
+    SpecialConsiderationPackage,
+    UbiPackage,
+)
 
 from .constants import PS_CONSTANTS_REPO_BRANCH, PS_CONSTANTS_REPO_URL
 from .core import (
     fetch_ps_constants,
     sync_compliance_priority,
+    sync_contract_priority,
     sync_special_consideration_packages,
     sync_ubi_packages,
 )
@@ -43,7 +49,12 @@ PS_CONSTANTS_BASE_URL = "/".join(
     # TODO: Use django_celery_beat which has PeriodicTask with IntervalSchedule
     #  What we use here is equivalent to PeriodicTask with CrontabSchedule
     crontab=crontab(minute="49", hour="*/5"),
-    data_models=[CompliancePriority, SpecialConsiderationPackage, UbiPackage],
+    data_models=[
+        CompliancePriority,
+        ContractPriority,
+        SpecialConsiderationPackage,
+        UbiPackage,
+    ],
 )
 def ps_constants_collector(collector_obj) -> str:
     """ps constants collector"""
@@ -52,6 +63,10 @@ def ps_constants_collector(collector_obj) -> str:
     url = "/".join((PS_CONSTANTS_BASE_URL, "compliance_priority.yml"))
     logger.info(f"Fetching PS Constants (compliance priority) from '{url}'")
     compliance_priority = fetch_ps_constants(url)
+
+    url = "/".join((PS_CONSTANTS_BASE_URL, "contract_priority.yml"))
+    logger.info(f"Fetching PS Constants (contract priority) from '{url}'")
+    contract_priority = fetch_ps_constants(url)
 
     url = "/".join((PS_CONSTANTS_BASE_URL, "ubi_packages.yml"))
     logger.info(f"Fetching PS Constants (Ubi Packages) from '{url}'")
@@ -71,6 +86,7 @@ def ps_constants_collector(collector_obj) -> str:
     )
 
     sync_compliance_priority(compliance_priority)
+    sync_contract_priority(contract_priority)
     sync_ubi_packages(ubi_packages)
     sync_special_consideration_packages(sc_packages)