+++ collectors/ps_constants/tests/test_core.py
@@ -3,12 +3,30 @@ import pytest
 from collectors.ps_constants.core import (
     fetch_ps_constants,
     sync_compliance_priority,
+    sync_contract_priority,
     sync_special_consideration_packages,
     sync_ubi_packages,
 )
-from osidb.models import CompliancePriority, SpecialConsiderationPackage, UbiPackage
+from osidb.models import (
+    CompliancePriority,
+    ContractPriority,
+    SpecialConsiderationPackage,
+    UbiPackage,
+)
 
 pytestmark = pytest.mark.unit
+
+
+SAMPLE_DATA = {
+    "rhel-1": [
+        "test-package",
+        "another-package",
+    ],
+    "rhel-2": [
+        "test-package",
+        "another-package",
+    ],
+}
 
 
 class TestPsConstantsCollection:
@@ -19,6 +37,11 @@ class TestPsConstantsCollection:
         compliance_priority = fetch_ps_constants(url)
         assert "rhel-9" in compliance_priority
         assert len(compliance_priority["rhel-9"]) > 0
+
+        url = "/".join((ps_constant_base_url, "contract_priority.yml"))
+        contract_priority = fetch_ps_constants(url)
+        assert "rhel-8" in contract_priority
+        assert len(contract_priority["rhel-8"]) > 0
 
         url = "/".join((ps_constant_base_url, "ubi_packages.yml"))
         ubi_packages = fetch_ps_constants(url)
@@ -33,23 +56,23 @@ class TestPsConstantsCollection:
         """
         test collector compliance priority data sync
         """
-        sampled_data = {
-            "rhel-1": [
-                "test-package",
-                "another-package",
-            ],
-            "rhel-2": [
-                "test-package",
-                "another-package",
-            ],
-        }
-        sync_compliance_priority(sampled_data)
+        sync_compliance_priority(SAMPLE_DATA)
 
         assert CompliancePriority.objects.all().count() == 4
         assert CompliancePriority.objects.filter(ps_module="rhel-1").count() == 2
         assert (
             CompliancePriority.objects.filter(ps_component="test-package").count() == 2
         )
+
+    def test_sync_contract_priority(self):
+        """
+        test collector contract priority data sync
+        """
+        sync_contract_priority(SAMPLE_DATA)
+
+        assert ContractPriority.objects.all().count() == 4
+        assert ContractPriority.objects.filter(ps_module="rhel-1").count() == 2
+        assert ContractPriority.objects.filter(ps_component="test-package").count() == 2
 
     def test_sync_ubi_packages(self):
         """Check collector can correctly sync ubi data in database"""